refactor: add all effects to animation.js for cleaner code
